refactor: improve admin dashboard and management modules for organizations, forum, and users (#132)
* refactor: improve admin dashboard and management modules for organizations, forum, and users

* refactor: improve comprehensive admin dashboard and management pages for users, mentorship, forum content, events, and audit logs

* refactor: improve comprehensive admin dashboard backend services and frontend management pages

* feat: add organization management components and admin dashboard layout pages

* refactor: change text to Vietnamese

* feat: implement AdminForumCategoriesPage and AdminForumPostsPage for managing forum structures and content

* refactor: change grid to flex

* feat: implement admin cross-organization article management APIs and update frontend endpoints

* feat: implement UpdateOrganizationRequest DTO and optimized SQL update query for organization management

* refactor: remove image upload functionality and update organization logo input to URL string field

* feat: extend organization introduction with vision, mission, core values, and banner fields in backend and frontend

* refactor: replace JSON configuration with a toggle-based UI for organization features

* feat: add status field to organization entity and update repository and service methods

* style: update table header background colors and improve breadcrumb navigation path detection

* feat: implement organization management module and add admin forum pages

* Delete backend/src/main/java/com/service/backend/config/DatabaseInitializer.java

---------

+++ backend/src/main/java/com/service/backend/organization/dao/OrganizationRepository.java
@@ -7,7 +7,7 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 @Repository
-public interface OrganizationRepository extends R2dbcRepository<Organization, Long> {
+public interface OrganizationRepository extends R2dbcRepository<Organization, Integer> {
 
     Mono<Organization> findBySlug(String slug);
